MP-202 :recycle: Refactor: Example 모듈을 Monitor 모듈로 재명명 및 각 객체 이름 수정
* 목적을 더욱 분명히 강조할 수 있는 이름 사용
